requirements.txt
numpy
torch
tqdm
pytest
pyyaml
tensorboardX
git+https://github.com/Hiroshiba/acoustic_feature_extractor
git+https://github.com/Hiroshiba/pytorch-trainer@1.1.0
